requirements.txt
xlrd
xlwt 
xlutils 
flask 
tqdm 
pandas 
camelot-py[cv] 
flask-migrate 
flask-sqlalchemy 
logger 
flask 
flask-login 
flask-bcrypt 
flask-login 
flask-wtf 
email-validator
flask-mail
jwt		
#camelot-py==0.10.1
#cffi==1.15.0
#chardet==4.0.0
#charset-normalizer==2.0.12
#click==8.1.2
#colorama==0.4.4
#cryptography==37.0.2
#et-xmlfile==1.1.0
#Flask==2.1.1
#ghostscript==0.7
#importlib-metadata==4.11.3
#itsdangerous==2.1.2
#Jinja2==3.1.1
#MarkupSafe==2.1.1
#numpy==1.22.3
#opencv-python==4.6.0.66
#openpyxl==3.0.10
#pandas==1.4.2
#pdfminer.six==20220524
#pdftopng==0.2.3
#pycparser==2.21
#PyPDF2==2.1.0
#python-dateutil==2.8.2
#pytz==2022.1
#six==1.16.0
#tabulate==0.8.9
#tqdm==4.64.0
#typing-extensions==4.2.0
#Werkzeug==2.1.1
#xlrd==0.7.1	
#xlutils==2.0.0
#xlwt==0.7.2
#zipp==3.8.0
#flask-sqlalchemy==0.7.7
